+++ acceptance/tests/vault/vault_namespaces_test.go
@@ -179,7 +179,6 @@ func TestVault_VaultNamespace(t *testing.T) {
 
 		"connectInject.enabled":  "true",
 		"connectInject.replicas": "1",
-		"controller.enabled":     "true",
 
 		"global.secretsBackend.vault.enabled":              "true",
 		"global.secretsBackend.vault.consulServerRole":     consulServerRole,